Update on the classes we made.
- We stayed by Integer instead of Float. (We discussed it with the visual team).
- XMLRead class now fills the worldObject list.
- Also fills the Transformation_matrix class.

+++ src/main/java/hu/oe/nik/szfmv/model/Classes/Tranzformacios_matrix.java
@@ -1,13 +1,13 @@
 package hu.oe.nik.szfmv.model.Classes;
 
 public class Tranzformacios_matrix {
-    float m11;
-    float m12;
-    float m21;
-    float m22;
+    int m11;
+    int m12;
+    int m21;
+    int m22;
 
 
-    public Tranzformacios_matrix(float m11, float m12, float m21, float m22) {
+    public Tranzformacios_matrix(int m11, int m12, int m21, int m22) {
         this.m11 = m11;
         this.m12 = m12;
         this.m21 = m21;
@@ -15,17 +15,17 @@ public class Tranzformacios_matrix {
     }
 
     public Tranzformacios_matrix(String m11s, String m12s, String m21s, String m22s) {
-        m11 = java.lang.Float.parseFloat(m11s);
-        m12 = java.lang.Float.parseFloat(m12s);
-        m21 = java.lang.Float.parseFloat(m21s);
-        m22 = java.lang.Float.parseFloat(m22s);
+        m11 = java.lang.Integer.parseInt(m11s);
+        m12 = java.lang.Integer.parseInt(m12s);
+        m21 = java.lang.Integer.parseInt(m21s);
+        m22 = java.lang.Integer.parseInt(m22s);
     }
 
     public float getM11() {
         return m11;
     }
 
-    public void setM11(float m11) {
+    public void setM11(int m11) {
         this.m11 = m11;
     }
 
@@ -33,7 +33,7 @@ public class Tranzformacios_matrix {
         return m12;
     }
 
-    public void setM12(float m12) {
+    public void setM12(int m12) {
         this.m12 = m12;
     }
 
@@ -41,7 +41,7 @@ public class Tranzformacios_matrix {
         return m21;
     }
 
-    public void setM21(float m21) {
+    public void setM21(int m21) {
         this.m21 = m21;
     }
 
@@ -49,7 +49,7 @@ public class Tranzformacios_matrix {
         return m22;
     }
 
-    public void setM22(float m22) {
+    public void setM22(int m22) {
         this.m22 = m22;
     }
 }
